Move JNM to own package and extracted TypeVariableContext
